+++ resources/views/partials/fields/horizontal.blade.php
@@ -1,6 +1,6 @@
 <div class="form-group row row-cols-sm-2 align-items-stretch">
     @if($field->has('title'))
-        <label for="{{ $field->get('id') }}" class="col-sm-3 text-wrap form-label">
+        <label for="{{ $field->get('id') }}" class="col-sm-3 text-wrap form-label text-balance">
             {{ $field->get('title') }}
 
             <x-orchid-popover :content="$field->get('popover', '')"/>
@@ -30,7 +30,7 @@
 
         @if($errors->has($field->getOldName()))
             <div class="invalid-feedback d-block">
-                <small>{{$errors->first($field->getOldName())}}</small>
+                <small>{{ $errors->first($field->getOldName()) }}</small>
             </div>
         @elseif($field->has('help'))
             <small class="form-text text-muted">{!! $field->get('help') !!}</small>