+++ resources/views/partials/fields/vertical.blade.php
@@ -1,6 +1,6 @@
 <div class="form-group">
     @if($field->has('title'))
-        <label for="{{ $field->get('id') }}" class="form-label">
+        <label for="{{ $field->get('id') }}" class="form-label text-balance">
             {{ $field->get('title') }}
 
             @if($field->get('required', false))
@@ -29,9 +29,9 @@
 
     @if($errors->has($field->getOldName()))
         <div class="invalid-feedback d-block">
-            <small>{{$errors->first($field->getOldName())}}</small>
+            <small>{{ $errors->first($field->getOldName()) }}</small>
         </div>
-    @elseif(isset($help))
+    @elseif($field->has('help'))
         <small class="form-text text-muted">
             {!! $field->get('help') !!}
         </small>