+++ src/Screen/Screen.php
@@ -267,7 +267,7 @@ abstract class Screen extends Controller
      */
     public function formValidateMessage(): string
     {
-        return __('Please check the entered data, it may be necessary to specify in other languages.');
+        return __('Unable to start action. Please check your input and try again.');
     }
 
     /**